turned off backward control for debugging

+++ MotorController.cpp
@@ -11,8 +11,8 @@
 
  void MotorController::control(StateEstimator * stateEstimator_p, velocity_setpoint_t * desiredVelocity_p, MotorDriver * driver_p)
  {
- 	int v = desiredVelocity_p->v;
- 	int w = desiredVelocity_p->w;
+ 	float v = desiredVelocity_p->v;
+ 	float w = desiredVelocity_p->w;
 
  	int rval = INV_MODEL_A*v + INV_MODEL_B*w;
  	int lval = INV_MODEL_C*v + INV_MODEL_D*w;
